alkjdsføalkdfjsa

+++ dat250-jpa-designproto/src/main/java/no/hvl/dat250/gruppe9/DAO/FeedPollResultDAO.java
@@ -19,7 +19,7 @@ public class FeedPollResultDAO {
         this.entityManager = entityManagerFactory.createEntityManager();
     }
 
-    public FeedPollResult getResult(int resultId){
+    public FeedPollResult getResult(long resultId){
         return entityManager.find(FeedPollResult.class, resultId);
     }
